backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlite3==2.6.0

# HTTP Clients
httpx==0.25.2

# Yandex Cloud Integration (ОСНОВНОЕ!)
requests==2.31.0

# Audio Processing (если используешь pydub)
pydub==0.25.1

# Google Docs (если нужно)
google-api-python-client==2.116.0
google-auth==2.41.1